Some AI Tweaks
The main thing was switching from TestCharacter to  "Asasasd" so that the AI no longer ignore the player

Also modified the EQS query as well to be more generous, in the hopes that if they walk into a wall they'll be able to turn around quicker. (Doesn't seem that effective)

Still need to add attack and switch off the debug model to a better one.