AI Attack
Rise of the machines.

The enemies now have randomised attack animations as well as idle animations.

I moved the enemy sounds out of the behaviour tree and into the AI controller, where it's called upon seeing the enemy.

Removing the need for the dedicated task.

This was done as the AI would wait on the task until the sound was played and the cooldown was finished.